move out gson logic
