+++ core/src/main/java/io/grpc/internal/ManagedChannelImpl.java
@@ -29,10 +29,6 @@ import com.google.common.base.Stopwatch;
 import com.google.common.base.Supplier;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonPrimitive;
 import io.grpc.Attributes;
 import io.grpc.CallOptions;
 import io.grpc.Channel;
@@ -55,7 +51,6 @@ import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
 import io.grpc.NameResolver;
 import io.grpc.Status;
-import io.grpc.Status.Code;
 import io.grpc.internal.Channelz.ChannelStats;
 import io.grpc.internal.ClientCallImpl.ClientTransportProvider;
 import io.grpc.internal.RetriableStream.ChannelBufferMeter;
@@ -68,10 +63,8 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -1093,100 +1086,12 @@ public final class ManagedChannelImpl extends ManagedChannel implements Instrume
     }
   }
 
+  // TODO(zdapeng): implement it once the Gson dependency issue is resolved.
   private static RetryPolicies getRetryPolicies(Attributes config) {
-    JsonObject serviceConfig = config.get(GrpcAttributes.NAME_RESOLVER_ATTR_SERVICE_CONFIG);
-    return getRetryPolicies(serviceConfig);
-  }
-
-  // TODO(zdapeng): maybe move it to a more appropriate place
-  @VisibleForTesting
-  static RetryPolicies getRetryPolicies(JsonObject serviceConfig) {
-    final Map<String, RetryPolicy> fullMethodNameMap = new HashMap<String, RetryPolicy>();
-    final Map<String, RetryPolicy> serviceNameMap = new HashMap<String, RetryPolicy>();
-
-    if (serviceConfig != null) {
-      JsonArray methodConfigs = serviceConfig.getAsJsonArray("methodConfig");
-
-      /* schema as follows
-      {
-        "methodConfig": [
-          {
-            "name": [
-              {
-                "service": string,
-                "method": string,         // Optional
-              }
-            ],
-            "retryPolicy": {
-              "maxAttempts": number,
-              "initialBackoff": string,   // Long decimal with "s" appended
-              "maxBackoff": string,       // Long decimal with "s" appended
-              "backoffMultiplier": number
-              "retryableStatusCodes": []
-            }
-          }
-        ]
-      }
-      */
-
-      if (methodConfigs != null) {
-        for (JsonElement methodConfig : methodConfigs) {
-          JsonArray names = methodConfig.getAsJsonObject().getAsJsonArray("name");
-          JsonObject retryPolicy = methodConfig.getAsJsonObject().getAsJsonObject("retryPolicy");
-          if (retryPolicy != null) {
-            int maxAttempts = retryPolicy.getAsJsonPrimitive("maxAttempts").getAsInt();
-            String initialBackoffStr =
-                retryPolicy.getAsJsonPrimitive("initialBackoff").getAsString();
-            checkState(
-                initialBackoffStr.charAt(initialBackoffStr.length() - 1) == 's',
-                "invalid value of initialBackoff");
-            double initialBackoff =
-                Double.parseDouble(initialBackoffStr.substring(0, initialBackoffStr.length() - 1));
-            String maxBackoffStr =
-                retryPolicy.getAsJsonPrimitive("maxBackoff").getAsString();
-            checkState(
-                maxBackoffStr.charAt(maxBackoffStr.length() - 1) == 's',
-                "invalid value of maxBackoff");
-            double maxBackoff =
-                Double.parseDouble(maxBackoffStr.substring(0, maxBackoffStr.length() - 1));
-            double backoffMultiplier =
-                retryPolicy.getAsJsonPrimitive("backoffMultiplier").getAsDouble();
-            JsonArray retryableStatusCodes = retryPolicy.getAsJsonArray("retryableStatusCodes");
-            Set<Code> codeSet = new HashSet<Code>(retryableStatusCodes.size());
-            for (JsonElement retryableStatusCode : retryableStatusCodes) {
-              codeSet.add(Code.valueOf(retryableStatusCode.getAsString()));
-            }
-            RetryPolicy pojoPolicy = new RetryPolicy(
-                maxAttempts, initialBackoff, maxBackoff, backoffMultiplier, codeSet);
-
-            for (JsonElement name : names) {
-              String service = name.getAsJsonObject().getAsJsonPrimitive("service").getAsString();
-              JsonPrimitive method = name.getAsJsonObject().getAsJsonPrimitive("method");
-              if (method != null && !method.getAsString().isEmpty()) {
-                fullMethodNameMap.put(
-                    MethodDescriptor.generateFullMethodName(service, method.getAsString()),
-                    pojoPolicy);
-              } else {
-                serviceNameMap.put(service, pojoPolicy);
-              }
-            }
-          }
-        }
-      }
-    }
-
     return new RetryPolicies() {
       @Override
       public RetryPolicy get(MethodDescriptor<?, ?> method) {
-        RetryPolicy retryPolicy = fullMethodNameMap.get(method.getFullMethodName());
-        if (retryPolicy == null) {
-          retryPolicy = serviceNameMap
-              .get(MethodDescriptor.extractFullServiceName(method.getFullMethodName()));
-        }
-        if (retryPolicy == null) {
-          retryPolicy = DEFAULT;
-        }
-        return retryPolicy;
+        return RetryPolicy.DEFAULT;
       }
     };
   }
